requirements.txt
git+ssh://git@github.com/kfirc/TelegramBot.git@v0.1
python-dotenv==0.14.0
python-dateutil==2.8.1
